adgm_corporate_agent/requirements.txt
gradio>=4.0.0
python-docx>=1.0.0
chromadb>=0.5.0
sentence-transformers>=2.2.2
torch>=2.2.0
numpy>=1.24.0
scikit-learn>=1.2.0
pdfminer.six>=20221105
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0
google-generativeai>=0.6.0
pytz>=2023.3
tqdm>=4.66.0
typing_extensions>=4.7.0
protobuf>=3.20.3,<6

